Update study_cohort_exclusion.sql
The goal with this edit is to avoid excluding kids who had a well child visit (WCV) listed in the session table. We only want to exclude kids who had a "relevant" session from 1/1/2017 through 5/31/2017 AND whose session was NOT a WCV. If a kid had a session that WAS a WCV during the same time period, we don't want to exclude them.